# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cricket Box Tournament API",
    "version": "1.0.0",
    "description": "Complete API documentation for Cricket Box - A cricket tournament management system with real-time match tracking, room management, and user statistics.",
    "contact": {
      "name": "Cricket Box Support",
      "email": "support@cricketbox.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.cricketbox.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "city": {
            "type": "string",
            "example": "Mumbai"
          },
          "role": {
            "type": "string",
            "enum": ["player", "admin", "umpire"],
            "example": "player"
          },
          "stats": {
            "type": "object",
            "properties": {
              "localMatches": {
                "type": "object",
                "properties": {
                  "matchesPlayed": { "type": "number", "example": 10 }
                }
              },
              "tournamentMatches": {
                "type": "object",
                "properties": {
                  "matchesPlayed": { "type": "number", "example": 5 }
                }
              }
            }
          },
          "totalStats": {
            "type": "object",
            "properties": {
              "matchesPlayed": { "type": "number", "example": 15 },
              "runs": { "type": "number", "example": 450 },
              "wickets": { "type": "number", "example": 12 }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "description": "Retrieve the authenticated user's profile information",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users (Admin only)",
        "description": "Retrieve a list of all registered users. Requires admin role.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
...
}